Reorganize API js
